Made persist dir and librarian dir configurable
The persist dir could be (and, indeed is) different from platform to platform,
so we made that configurable via kconfig. The librarian settings path is now
set using the $BR2_LIBRARIAN_SETTINGS_FILE variable. It is up to the build to
assign a value to this variable, though, and no menu items is provided.

Signed-off-by: Branko Vukelic <branko@outernet.is>