Fix circular dependency issue with librarian-*
Apparently, when the name of another package is included in the name of some
other package, the first package is automatically delcared a dependency of the
second pacakge. Trying to 'select' the dependent package results in circular
dependency. To address this with librarian-* packages, 'select' lines are
removed from librarian-analytics and librarian-netiterfaces.

Signed-off-by: Branko Vukelic <branko@outernet.is>