Add tarball install option to ONDD
This patch modifies the ondd package so that it has three install modes
selectable through the configuration menu. In addition to compiling from
soruce, and simply not installing, the third option to install the binary from
a hosted tarball has been added.

Due to the multitude of install options, the patches are now moved to srcpatch
subdir, and are listed manually only for the build-from-source option.

The source directory can now also be specified through the configuration menu,
so there is no need to use OVERRIDE_SRCDIR anymore.

Signed-off-by: Branko Vukelic <branko@outernet.is>