Use -b and -m switches, clean up PID file
This patch fixes the init script so it uses -b switch instead of manually
backgrounding the proc, and also adds the -m switch so that the PID file is
actually written. In the stop section, the PID file is also cleaned up as
start-stop-daemon apparently won't do it for us.

Signed-off-by: Branko Vukelic <branko@outernet.is>